pyproject.toml
[project]
name = "pyapi-init"
version = "0.1.0"
description = "Production ready FastAPI init code."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiosqlite>=0.22.1",
  "alembic>=1.18.1",
  "argon2-cffi>=25.1.0",
  "asyncpg>=0.31.0",
  "celery>=5.6.2",
  "click>=8.3.1",
  "fastapi-mail>=1.6.1",
  "fastapi[standard]>=0.127.0",
  "passlib[argon2]>=1.7.4",
  "premailer>=3.10.0",
  "psycopg[binary]>=3.3.2",
  "pydantic-settings>=2.12.0",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.21",
  "redis>=7.1.0",
  "slowapi>=0.1.9",
  "sqlmodel>=0.0.29",
  "uuid6>=2025.0.1",
  "uvicorn[standard]>=0.40.0",
]

# =============================================================================
# Development Dependencies.
# =============================================================================

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "black>=25.12.0",
  "isort>=7.0.0",
  "types-python-jose>=3.5.0.20250531",
  "celery-types>=0.24.0",
  "types-passlib>=1.7.7.20260211",
]

# =============================================================================
# Pyright Configurations.
# =============================================================================

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUntypedFunctionDecorator = false

# =============================================================================
# Disable Ruff Rules for Test Cases.
# =============================================================================

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

# =============================================================================
# Pytest Configurations.
# =============================================================================

[tool.pytest.ini_options]
minversion = "8.0"
addopts = """
--strict-markers
--strict-config
--tb=short
--cov=app
--cov-report=term-missing
--cov-report=html
--cov-report=xml
--no-cov-on-fail
"""
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
filterwarnings = ["error"]
markers = [
  "unit: Unit tests",
  "integration: Integration tests",
  "e2e: End-to-end tests",
  "slow: Slow tests",
]

# =============================================================================
# Pytest Coverage Configurations.
# =============================================================================

[tool.coverage.run]
branch = true
source = ["app"]

# =============================================================================
# Pytest Coverage Report Configurations.
# =============================================================================

[tool.coverage.report]
# fail_under = 85
show_missing = true
skip_covered = true
